Add test for editing card

+++ src/server/services/card/hooks/removeMp3.js
@@ -6,6 +6,10 @@ AWS.config.update({
 const s3 = new AWS.S3()
 
 module.exports = async filename => {
+    if (process.env.NODE_ENV !== 'production') {
+        return
+    }
+
     await s3
         .deleteObject({
             Bucket: process.env.AWS_S3_BUCKET,
